SCHOOL-3 finished factory and seeder for dummy data

+++ database/migrations/2023_08_07_043249_create_exams_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->integer('type');
             $table->unsignedBigInteger('subject_id');
             $table->unsignedBigInteger('student_id');
-            $table->integer('score')->nullable();
+            $table->double('score')->nullable();
             $table->string('review')->nullable();
             $table->timestamps();
             $table->timestamp('deleted_at')->nullable();
